Extract ProofreadingPassage model from proofreadings controller and ProofreaderActivity
